+++ src/screens/RankingScreen.js
@@ -1,66 +1,8 @@
 import React from 'react';
-import {
-  View,
-  StyleSheet,
-  Image,
-  TouchableOpacity,
-  Text,
-  Dimensions,
-} from 'react-native';
-import { KeyboardAwareScrollView } from 'react-native-keyboard-aware-scroll-view';
+import { View, StyleSheet, Image } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import { Button, Appbar, DataTable, Card } from 'react-native-paper';
-import theme from '../core/theme';
-
-const screenWidth = Dimensions.get('screen').width;
-const screenHeight = Dimensions.get('screen').height;
-
-const styles = StyleSheet.create({
-  view: {
-    flex: 1,
-    width: '100%',
-    justifyContent: 'center',
-  },
-  appbar: {
-    backgroundColor: theme.colors.background,
-    height: 100,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  image: {
-    height: 120,
-    position: 'relative',
-    marginBottom: 12,
-    marginTop: 100,
-    width: 128,
-    alignItems: 'center',
-    zIndex: 2,
-  },
-  table: {
-    backgroundColor: theme.colors.background,
-    width: 350,
-    marginTop: 70,
-    marginLeft: 35,
-    justifyContent: 'center',
-  },
-  ranking: {
-    justifyContent: 'space-between',
-  },
-  button: {
-    backgroundColor: theme.colors.secondary,
-    borderRadius: 20,
-    height: 45,
-    justifyContent: 'center',
-    marginVertical: 20,
-    marginLeft: 90,
-    width: 230,
-  },
-  buttontext: {
-    fontSize: 15,
-    fontWeight: 'bold',
-    justifyContent: 'center',
-  },
-});
+import { Button, Appbar, DataTable } from 'react-native-paper';
+import styles from '../styles/RankingScreenStyle';
 
 const RankingScreen = () => {
   const navigation = useNavigation();
